Added module view_model, renamed some classes, refactored gradles...
