+++ compose_app/build.gradle
@@ -70,6 +70,9 @@ dependencies {
     // Coil for Image Loading
     implementation("io.coil-kt:coil-compose:1.3.1")
 
+    // Work Manager
+    implementation "androidx.work:work-runtime-ktx:$work_version"
+
     // Koin version
     // Koin for Dependency Injection
     implementation "io.insert-koin:koin-core:$koin_version"
