+++ settings.gradle
@@ -4,3 +4,4 @@ include ':data'
 include ':ui'
 include ':compose_app'
 include ':compose_ui'
+include ':view_model'
