+++ ui/build.gradle
@@ -4,11 +4,11 @@ plugins {
 }
 
 android {
-    compileSdk 30
+    compileSdk 31
 
     defaultConfig {
         minSdk 26
-        targetSdk 30
+        targetSdk 31
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles "consumer-rules.pro"
@@ -36,6 +36,7 @@ android {
 
 dependencies {
     implementation project(':domain')
+    implementation project(':view_model')
 
     // UnitTest
     testImplementation "junit:junit:$unit_test_version"
@@ -46,9 +47,6 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:$lifecycle_livedata_ktx_version"
     implementation "androidx.cardview:cardview:$cardview_version"
     implementation "com.google.android.material:material:$material_version"
-
-    // ViewModel
-    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_viewmodel_ktx_version"
 
     // Fragments
     implementation "androidx.fragment:fragment-ktx:$fragments_ktx_version"
