+++ view_model/src/main/java/com/view_model/NewsListViewModel.kt
@@ -1,28 +1,25 @@
-package ui.viewModel
+package com.view_model
 
 import android.app.Application
 import android.content.Context
 import android.net.ConnectivityManager
-import android.util.Log
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
-import domain.NewsItemDetails
-import domain.NewsRepository
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.launch
+import com.view_model.loadState.LoadState
+import domain.NewsItem
+import kotlinx.coroutines.*
+import kotlinx.coroutines.flow.*
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
+import domain.NewsRepository
 import java.io.IOException
 
-class NewsPageViewModel(application: Application,
-                        itemID: String)
+class NewsListViewModel (application: Application)
     : AndroidViewModel(application), KoinComponent {
 
     private val newsRepository: NewsRepository by inject()
 
-    val newsPageFlow: Flow<NewsItemDetails> = newsRepository.getItemDetailsByID(itemID)
+    val newsListFlow: Flow<List<NewsItem>> = newsRepository.getItems()
 
     val state: MutableStateFlow<LoadState> = MutableStateFlow(LoadState.IDLE)
 
@@ -33,11 +30,11 @@ class NewsPageViewModel(application: Application,
         return  networkInfo!=null && networkInfo.isConnected
     }
 
-    fun loadData(itemID: String) {
-        viewModelScope.launch {
+    fun loadData() {
+        viewModelScope.launch(Dispatchers.Main) {
             state.value = LoadState.LOADING
             try {
-                newsRepository.loadNewsItemDetailsByID(itemID)
+                newsRepository.loadNews()
                 state.value = LoadState.SUCCESS
             } catch (e: Throwable) {
                 if (!isConnectedToInternet() && e is IOException) {
@@ -45,10 +42,10 @@ class NewsPageViewModel(application: Application,
                 }
                 else if (e is NullPointerException) {
                     when (e.message) {
+                        "Error: List<NewsItem> from json is empty!" ->
+                            state.value = LoadState.EMPTY_ITEMS_LIST_ERROR
                         "Error: List<NewsItemDetails> from json is empty!" ->
                             state.value = LoadState.EMPTY_ITEMS_DETAILS_LIST_ERROR
-                        "Error: It was unable to find out the item with chosen id!" ->
-                            state.value = LoadState.INCONSISTENCY_ITEM_ID_ERROR
                     }
                 }
                 else {
@@ -56,11 +53,6 @@ class NewsPageViewModel(application: Application,
                 }
             }
         }
-    }
-
-    override fun onCleared() {
-        super.onCleared()
-        Log.i("SportViewModel", "SportViewModel destroyed!")
     }
 
 }
