+++ view_model/src/main/java/com/view_model/NewsPageViewModel.kt
@@ -1,24 +1,28 @@
-package ui.viewModel
+package com.view_model
 
 import android.app.Application
 import android.content.Context
 import android.net.ConnectivityManager
+import android.util.Log
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
-import domain.NewsItem
-import kotlinx.coroutines.*
-import kotlinx.coroutines.flow.*
+import com.view_model.loadState.LoadState
+import domain.NewsItemDetails
+import domain.NewsRepository
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.launch
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
-import domain.NewsRepository
 import java.io.IOException
 
-class NewsListViewModel (application: Application)
+class NewsPageViewModel(application: Application,
+                        itemID: String)
     : AndroidViewModel(application), KoinComponent {
 
     private val newsRepository: NewsRepository by inject()
 
-    val newsListFlow: Flow<List<NewsItem>> = newsRepository.getItems()
+    val newsPageFlow: Flow<NewsItemDetails> = newsRepository.getItemDetailsByID(itemID)
 
     val state: MutableStateFlow<LoadState> = MutableStateFlow(LoadState.IDLE)
 
@@ -29,11 +33,11 @@ class NewsListViewModel (application: Application)
         return  networkInfo!=null && networkInfo.isConnected
     }
 
-    fun loadData() {
-        viewModelScope.launch(Dispatchers.Main) {
+    fun loadData(itemID: String) {
+        viewModelScope.launch {
             state.value = LoadState.LOADING
             try {
-                newsRepository.loadNews()
+                newsRepository.loadNewsItemDetailsByID(itemID)
                 state.value = LoadState.SUCCESS
             } catch (e: Throwable) {
                 if (!isConnectedToInternet() && e is IOException) {
@@ -41,10 +45,10 @@ class NewsListViewModel (application: Application)
                 }
                 else if (e is NullPointerException) {
                     when (e.message) {
-                        "Error: List<NewsItem> from json is empty!" ->
-                            state.value = LoadState.EMPTY_ITEMS_LIST_ERROR
                         "Error: List<NewsItemDetails> from json is empty!" ->
                             state.value = LoadState.EMPTY_ITEMS_DETAILS_LIST_ERROR
+                        "Error: It was unable to find out the item with chosen id!" ->
+                            state.value = LoadState.INCONSISTENCY_ITEM_ID_ERROR
                     }
                 }
                 else {
@@ -52,6 +56,11 @@ class NewsListViewModel (application: Application)
                 }
             }
         }
+    }
+
+    override fun onCleared() {
+        super.onCleared()
+        Log.i("SportViewModel", "SportViewModel destroyed!")
     }
 
 }
