+++ view_model/src/main/java/com/view_model/loadState/LoadState.kt
@@ -1,6 +1,5 @@
-package ui.viewModel
+package com.view_model.loadState
 
-//TODO: move to ui module
 enum class LoadState {
     SUCCESS,
     LOADING,
